Move Mermaid History stuff to Caravel.History.Mermaid and Graph stuff to Caravel.DijkstraGraph to allow custom implementations.
