+++ Caravel.sln
@@ -28,6 +28,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caravel.Tests.Fixtures", "test\Caravel.Tests.Fixtures\Caravel.Tests.Fixtures.csproj", "{2E70A971-648E-4D93-A717-8B6974160A3C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caravel.History.Mermaid", "src\Caravel.History.Mermaid\Caravel.History.Mermaid.csproj", "{ECD3FF43-3F8F-419F-9EDF-2884D49AD758}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Caravel.DijkstraGraph", "src\Caravel.DijkstraGraph\Caravel.DijkstraGraph.csproj", "{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -54,6 +58,14 @@ Global
 		{2E70A971-648E-4D93-A717-8B6974160A3C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E70A971-648E-4D93-A717-8B6974160A3C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2E70A971-648E-4D93-A717-8B6974160A3C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ECD3FF43-3F8F-419F-9EDF-2884D49AD758}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ECD3FF43-3F8F-419F-9EDF-2884D49AD758}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ECD3FF43-3F8F-419F-9EDF-2884D49AD758}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ECD3FF43-3F8F-419F-9EDF-2884D49AD758}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -65,6 +77,8 @@ Global
 		{C1002264-B5E4-4F48-B542-717A11A55CD2} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{E9540A45-EA02-4F9F-8C57-D83AA246CFE4} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{2E70A971-648E-4D93-A717-8B6974160A3C} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{ECD3FF43-3F8F-419F-9EDF-2884D49AD758} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{50CC11EE-3B10-4AA5-A161-AAACF0DA8DAC} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C117A897-A1ED-4783-8937-F9378E58B728}
