+++ src/Caravel.DijkstraGraph/Caravel.DijkstraGraph.csproj
@@ -4,6 +4,7 @@
     </PropertyGroup>
     <ItemGroup>
         <ProjectReference Include="..\Caravel.Abstractions\Caravel.Abstractions.csproj" />
-        <ProjectReference Include="..\Caravel.History.Mermaid\Caravel.History.Mermaid.csproj" />
+        <ProjectReference Include="..\Caravel.Core\Caravel.Core.csproj" />
     </ItemGroup>
 </Project>
+
