+++ src/Caravel.DijkstraGraph/Extensions/GraphExtensions.cs
@@ -1,0 +1,40 @@
+﻿using Caravel.Abstractions;
+using Caravel.Core;
+
+namespace Caravel.DijkstraGraph.Extensions;
+
+public static class GraphExtensions
+{
+    public static IRoute GetShortestRoute(this IGraph graph, Type origin, Type destination)
+    {
+        ArgumentNullException.ThrowIfNull(graph, nameof(graph));
+        return graph.GetShortestRoute(
+            origin,
+            destination,
+            Waypoints.Empty(),
+            ExcludedNodes.Empty()
+        );
+    }
+
+    public static IRoute GetShortestRoute(
+        this IGraph graph,
+        Type origin,
+        Type destination,
+        IExcludedNodes excludedNodes
+    )
+    {
+        ArgumentNullException.ThrowIfNull(graph, nameof(graph));
+        return graph.GetShortestRoute(origin, destination, Waypoints.Empty(), excludedNodes);
+    }
+
+    public static IRoute GetShortestRoute(
+        this IGraph graph,
+        Type origin,
+        Type destination,
+        IWaypoints waypoints
+    )
+    {
+        ArgumentNullException.ThrowIfNull(graph, nameof(graph));
+        return graph.GetShortestRoute(origin, destination, waypoints, ExcludedNodes.Empty());
+    }
+}
