+++ src/Caravel.DijkstraGraph/GlobalSuppressions.cs
