+++ src/Caravel.History.Mermaid/GraphExtensions.cs
@@ -1,11 +1,48 @@
-﻿using System.Globalization;
+﻿using System.Collections.Immutable;
+using System.Globalization;
 using System.Text;
 using Caravel.Abstractions;
 
-namespace Caravel.Core.Extensions;
+namespace Caravel.History.Mermaid;
 
-public static class MermaidGraphExtensions
+public static class GraphExtensions
 {
+    public static string ToMermaidGraph(
+        this IGraph graph,
+        MermaidGraphDirections mermaidGraphDirection = default
+    )
+    {
+        ArgumentNullException.ThrowIfNull(graph, nameof(graph));
+        var stringBuilder = new StringBuilder();
+        stringBuilder.AppendLine(
+            CultureInfo.InvariantCulture,
+            $"graph {mermaidGraphDirection.ToString()}"
+        );
+
+        // order need to be respected for comparison
+        List<(Type node, ImmutableHashSet<IEdge> edges)> nodeEdges =
+        [
+            .. graph
+                .Nodes.OrderBy(x => x.Key.Name)
+                .Select(n => (node: n.Key, edges: n.Value.GetEdges())),
+        ];
+
+        for (var i = 0; i < nodeEdges.Count; i++)
+        {
+            foreach (var edge in nodeEdges[i].edges.OrderBy(x => x.Neighbor.Name))
+            {
+                var edgeStr = edge.ToString();
+                stringBuilder.AppendLine(edgeStr);
+            }
+        }
+
+        var result = stringBuilder.ToString();
+        return result.EndsWith(Environment.NewLine, StringComparison.OrdinalIgnoreCase)
+            ? result[..^Environment.NewLine.Length]
+            : result;
+        ;
+    }
+
     public static string ToMermaidHtml(
         this IGraph graph,
         MermaidGraphDirections mermaidGraphDirection = default
