+++ src/Caravel.History.Mermaid/JourneyHistory.cs
@@ -1,6 +1,6 @@
 ﻿using Caravel.Abstractions;
 
-namespace Caravel.Core;
+namespace Caravel.History.Mermaid;
 public record JourneyHistory : IJourneyHistory
 {
     public Queue<IEdge> Edges { get; init; } = new Queue<IEdge>();
