+++ test/Caravel.Tests.Fixtures/Caravel.Tests.Fixtures.csproj
@@ -16,5 +16,7 @@
     <ItemGroup>
         <ProjectReference Include="..\..\src\Caravel.Abstractions\Caravel.Abstractions.csproj" />
         <ProjectReference Include="..\..\src\Caravel.Core\Caravel.Core.csproj" />
+        <ProjectReference Include="..\..\src\Caravel.DijkstraGraph\Caravel.DijkstraGraph.csproj" />
+        <ProjectReference Include="..\..\src\Caravel.History.Mermaid\Caravel.History.Mermaid.csproj" />
     </ItemGroup>
 </Project>
