+++ test/Caravel.Tests.Fixtures/GraphTestBuilder.cs
@@ -1,5 +1,6 @@
 ﻿using Caravel.Abstractions;
 using Caravel.Core;
+using Caravel.DijkstraGraph;
 using NSubstitute;
 
 namespace Caravel.Tests.Fixtures;
