+++ test/UnitTests/Caravel.Core.UnitTests/Caravel.Core.UnitTests.csproj
@@ -23,7 +23,10 @@
         <PackageReference Include="AwesomeAssertions" />
     </ItemGroup>
     <ItemGroup>
+        <ProjectReference Include="..\..\..\src\Caravel.Abstractions\Caravel.Abstractions.csproj" />
         <ProjectReference Include="..\..\..\src\Caravel.Core\Caravel.Core.csproj" />
+        <ProjectReference Include="..\..\..\src\Caravel.DijkstraGraph\Caravel.DijkstraGraph.csproj" />
+        <ProjectReference Include="..\..\..\src\Caravel.History.Mermaid\Caravel.History.Mermaid.csproj" />
         <ProjectReference Include="..\..\Caravel.Tests.Fixtures\Caravel.Tests.Fixtures.csproj" />
         <ProjectReference Include="..\..\Caravel.Tests.Support\Caravel.Tests.Support.csproj" />
     </ItemGroup>
