+++ test/UnitTests/Caravel.Core.UnitTests/GlobalUsing.cs
@@ -1,9 +1,12 @@
-﻿global using AwesomeAssertions;
-global using Caravel.Core.Extensions;
+﻿global using Caravel.Core.Extensions;
 global using Caravel.Core.UnitTests.Extensions;
 global using Caravel.Tests.Fixtures;
 global using Caravel.Tests.Fixtures.NodeSpies;
 global using Xunit;
 global using static Caravel.Core.UnitTests.Extensions.VerifyExtensions;
+global using static Caravel.History.Mermaid.GraphExtensions;
+global using static Caravel.History.Mermaid.JourneyExtensions;
+global using static Caravel.History.Mermaid.RouteExtensions;
 global using static Caravel.Tests.Support.FeaturesUnderTest;
 global using static Caravel.Tests.Support.TestTypes;
+
